requirements.txt
Flask>=2.0
rapidfuzz>=3.0
python-dotenv>=1.0
apidfuzz 
jellyfish
requests 
beautifulsoup4 
python-dotenv
bs4